# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Checklist API",
    "version": "1.0"
  },
  "tags": [
    {
      "name": "templates",
      "description": "Endpoints for managing templates"
    },
    {
      "name": "checklists",
      "description": "Endpoints for managing checklists"
    }
  ],
  "paths": {
    "/templates": {
      "get": {
        "tags": ["templates"],
        "summary": "Get all templates and their nodes",
        "description": "Retrieves all templates and their nodes from the database.",
        "responses": {
          "200": {
            "description": "Templates found",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TemplateNodes"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["templates"],
        "summary": "Create a new template along with template nodes",
        "description": "Creates a new template and its nodes in the database.",
        "parameters": [
          {
            "name": "template",
            "in": "body",
            "description": "Template to create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TemplateNodes"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Template created",
            "schema": {
              "$ref": "#/definitions/TemplateNodes"
            }
          }
        }
      }
    },
    "/templates/{id}": {
      "get": {
        "tags": ["templates"],
        "summary": "Get a template and its nodes by template_ID",
        "description": "Retrieves a template from the database by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the template to retrieve",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Template found",
            "schema": {
              "$ref": "#/definitions/TemplateNodes"
            }
          },
          "404": {
            "description": "Template not found"
          }
        }
      },
      "put": {
        "tags": ["templates"],
        "summary": "Update a template and its nodes",
        "description": "Updates an existing template in the database.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the template to retrieve",
            "required": true,
            "type": "string"
          },
          {
            "name": "template",
            "in": "body",
            "description": "Template data to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TemplateNodes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Template updated",
            "schema": {
              "$ref": "#/definitions/Template"
            }
          }
        }
      },
      "delete": {
        "tags": ["templates"],
        "summary": "Delete a template",
        "description": "Deletes an existing template from the database.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the template to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Template deleted",
            "schema": {
              "$ref": "#/definitions/Template"
            }
          }
        }
      }
    },
    "/checklists": {
      "get": {
        "tags": ["checklists"],
        "summary": "Get all checklists and their nodes",
        "description": "Retrieves all checklists and their nodes from the database.",
        "responses": {
          "200": {
            "description": "checklists found",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ChecklistNodes"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["checklists"],
        "summary": "Create a new checklist along with checklist nodes",
        "description": "Creates a new checklist and its nodes in the database.",
        "parameters": [
          {
            "name": "checklist",
            "in": "body",
            "description": "checklist to create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChecklistNodes"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Template created",
            "schema": {
              "$ref": "#/definitions/ChecklistNodes"
            }
          }
        }
      }
    },
    "/checklists/{id}": {
      "get": {
        "tags": ["checklists"],
        "summary": "Get a checklist and its nodes by checklist_ID",
        "description": "Retrieves a checklist from the database by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the checklist to retrieve",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "checklist found",
            "schema": {
              "$ref": "#/definitions/ChecklistNodes"
            }
          },
          "404": {
            "description": "checklist not found"
          }
        }
      },
      "put": {
        "tags": ["checklists"],
        "summary": "Update a checklist and its nodes",
        "description": "Updates an existing checklist in the database.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the checklist to retrieve",
            "required": true,
            "type": "string"
          },
          {
            "name": "checklist",
            "in": "body",
            "description": "checklist data to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChecklistNodes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "checklist updated",
            "schema": {
              "$ref": "#/definitions/Checklist"
            }
          }
        }
      },
      "delete": {
        "tags": ["checklists"],
        "summary": "Delete a checklist",
        "description": "Deletes an existing checklist from the database.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the checklist to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "checklist deleted",
            "schema": {
              "$ref": "#/definitions/Checklist"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Template": {
      "type": "object",
      "properties": {
        "template_id": {
          "type": "string",
          "example": "7d9315a0-fa47-482e-85d8-1d48443fc538"
        },
        "template_title": {
          "type": "string",
          "example": "Flight Towing List"
        },
        "template_owner_id": {
          "type": "string",
          "example": "akhilesh"
        },
        "template_created_at":{
          "type":"string",
          "format":"date-time"
        }
      }
    },
    "TemplateNodes": {
      "type": "object",
      "properties": {
        "template_id": {
          "type": "string",
          "example": "12345"
        },
        "template_title": {
          "type": "string",
          "example": "Flight Towing List"
        },
        "template_owner_id": {
          "type": "string",
          "example": "akhilesh"
        },
        "template_created_at":{
          "type":"string",
          "format":"date-time"
        },
        "template_nodes":{
          "type":"array",
          "items":{
            "$ref":"#/definitions/TemplateNode"
          }
        }
      }
    },
    "TemplateNode":{
      "type":"object",
      "properties":{
        "node_id":{
          "type":"string",
          "example":"123321"
        },
        "node_title":{
          "type":"string",
          "example":"Caution Title"
        },
        "node_description":{
          "type":"string",
          "example":"<p>Enter description for this caution.</p>"
        },
        "node_type":{
          "type":"string",
          "example":"Caution"
        },
        "template_id":{
          "type":"string",
          "example":"12345"
        },
        "node_index":{
          "type":"integer",
          "example":1
        }
      }
    },
    "ChecklistNodes": {
      "type": "object",
      "properties": {
        "checklist_id": {
          "type": "string",
          "example": "12345"
        },
        "checklist_title": {
          "type": "string",
          "example": "Flight Towing List"
        },
        "checklist_created_at":{
          "type":"string",
          "format":"date-time"
        },
        "checklist_nodes":{
          "type":"array",
          "items":{
            "$ref":"#/definitions/ChecklistNode"
          }
        },
        "user_id": {
          "type": "string",
          "example": "Manager"
        },
        "checklist_completion": {
          "type": "boolean",
          "example": false
        },
        "checklist_completion_timestamp": {
          "type":"string",
          "format":"date-time"
        }
      }
    },
    "ChecklistNode":{
      "type":"object",
      "properties":{
        "checklist_node_id":{
          "type":"string",
          "example":"123321"
        },
        "checklist_id":{
          "type":"string",
          "example":"12345"
        },
        "checklist_node_title":{
          "type":"string",
          "example":"Caution Title"
        },
        "checklist_node_description":{
          "type":"string",
          "example":"<p>Enter description for this caution.</p>"
        },
        "checklist_node_type":{
          "type":"string",
          "example":"Caution"
        },
        "checklist_node_completion":{
          "type":"boolean",
          "example":false
        },
        "checklist_node_completion_timestamp":{
          "type":"string",
          "format":"date-time"
        },
        "checklist_node_index":{
          "type":"integer",
          "example":1
        }
      }
    }
  }
}